Login/Create Account
Creates the jsx and css for the Login and Create Account. Additionally, I made some modification to the nav bar and app .jsx. Changes were minor. Adding lines of code to redirect the user to the login and create account. Finally, I had to install something to the packahe files such as         "mdb-react-ui-kit": "^9.0.0". And the last thing I did was upload the dsa logo to the image files.